+++ generators/typescript/model/type-reference-converters/src/TypeReferenceToStringExpressionConverter.ts
@@ -83,7 +83,12 @@ export class TypeReferenceToStringExpressionConverter extends AbstractTypeRefere
                                 nullable: (nullable) => getStringify(this.context.type.resolveTypeReference(nullable)),
                                 set: () => this.jsonStringify(mapExpression),
                                 map: () => this.jsonStringify(mapExpression),
-                                literal: () => this.jsonStringifyIfNotString(mapExpression),
+                                literal: (literal) => {
+                                    if (literal.type === "string") {
+                                        return mapExpression;
+                                    }
+                                    return this.jsonStringify(mapExpression);
+                                },
                                 _other: () => {
                                     throw new Error("Unknown ContainerType: " + containerType.type);
                                 }